+++ packages/walletkit/src/evm/wallets/index.ts
@@ -8,7 +8,7 @@ export * from './trustWallet';
 export * from './walletConnect';
 export * from './okxWallet';
 export * from './mathWallet';
-export * from './binanceWeb3Wallet';
+export * from './binanceWallet';
 export * from './coinbaseWallet';
 export * from './bitgetWallet';
 export * from './safe';